refactor(cz/utils): rename filters as utils
